+++ packages/@uppy/aws-s3-multipart/src/MultipartUploader.js
@@ -24,26 +24,47 @@ function ensureInt (value) {
   throw new TypeError('Expected a number')
 }
 
-const pausingUploadReason = Symbol('pausing upload, not an actual error')
+export const pausingUploadReason = Symbol('pausing upload, not an actual error')
 
+/**
+ * A MultipartUploader instance is used per file upload to determine whether a
+ * upload should be done as multipart or as a regular S3 upload
+ * (based on the user-provided `shouldUseMultipart` option value) and to manage
+ * the chunk splitting.
+ */
 class MultipartUploader {
   #abortController = new AbortController()
 
+  /** @type {import("../types/chunk").Chunk[]} */
   #chunks
 
+  /** @type {{ uploaded: number, etag?: string, done?: boolean }[]} */
   #chunkState
 
+  /**
+   * The (un-chunked) data to upload.
+   *
+   * @type {Blob}
+   */
   #data
 
+  /** @type {import("@uppy/core").UppyFile} */
   #file
 
-  #uploadPromise
+  /** @type {boolean} */
+  #uploadHasStarted = false
 
+  /** @type {(err?: Error | any) => void} */
   #onError
 
+  /** @type {() => void} */
   #onSuccess
 
+  /** @type {import('../types/index').AwsS3MultipartOptions["shouldUseMultipart"]} */
   #shouldUseMultipart
+
+  /** @type {boolean} */
+  #isRestoring
 
   #onReject = (err) => (err?.cause === pausingUploadReason ? null : this.#onError(err))
 
@@ -64,6 +85,11 @@ class MultipartUploader {
     this.#onSuccess = this.options.onSuccess
     this.#onError = this.options.onError
     this.#shouldUseMultipart = this.options.shouldUseMultipart
+
+    // When we are restoring an upload, we already have an UploadId and a Key. Otherwise
+    // we need to call `createMultipartUpload` to get an `uploadId` and a `key`.
+    // Non-multipart uploads are not restorable.
+    this.#isRestoring = options.uploadId && options.key
 
     this.#initChunks()
   }
@@ -90,12 +116,12 @@ class MultipartUploader {
       }
       this.#chunks = Array(arraySize)
 
-      for (let i = 0, j = 0; i < fileSize; i += chunkSize, j++) {
-        const end = Math.min(fileSize, i + chunkSize)
+      for (let offset = 0, j = 0; offset < fileSize; offset += chunkSize, j++) {
+        const end = Math.min(fileSize, offset + chunkSize)
 
         // Defer data fetching/slicing until we actually need the data, because it's slow if we have a lot of files
         const getData = () => {
-          const i2 = i
+          const i2 = offset
           return this.#data.slice(i2, end)
         }
 
@@ -104,6 +130,14 @@ class MultipartUploader {
           onProgress: this.#onPartProgress(j),
           onComplete: this.#onPartComplete(j),
           shouldUseMultipart,
+        }
+        if (this.#isRestoring) {
+          const size = offset + chunkSize > fileSize ? fileSize - offset : chunkSize
+          // setAsUploaded is called by listPart, to keep up-to-date the
+          // quantity of data that is left to actually upload.
+          this.#chunks[j].setAsUploaded = () => {
+            this.#chunkState[j].uploaded = size
+          }
         }
       }
     } else {
@@ -119,13 +153,14 @@ class MultipartUploader {
   }
 
   #createUpload () {
-    this.#uploadPromise = this
+    this
       .options.companionComm.uploadFile(this.#file, this.#chunks, this.#abortController.signal)
       .then(this.#onSuccess, this.#onReject)
+    this.#uploadHasStarted = true
   }
 
   #resumeUpload () {
-    this.#uploadPromise = this
+    this
       .options.companionComm.resumeUploadFile(this.#file, this.#chunks, this.#abortController.signal)
       .then(this.#onSuccess, this.#onReject)
   }
@@ -158,9 +193,12 @@ class MultipartUploader {
   }
 
   start () {
-    if (this.#uploadPromise) {
+    if (this.#uploadHasStarted) {
       if (!this.#abortController.signal.aborted) this.#abortController.abort(pausingUploadReason)
       this.#abortController = new AbortController()
+      this.#resumeUpload()
+    } else if (this.#isRestoring) {
+      this.options.companionComm.restoreUploadFile(this.#file, { uploadId: this.options.uploadId, key: this.options.key })
       this.#resumeUpload()
     } else {
       this.#createUpload()